fix: remove unused code and retrigger on change
