+++ src/e2e.ts
@@ -1,8 +1,10 @@
 import { ofetch } from 'ofetch'
 import type { FetchOptions, FetchResponse, MappedResponseType, ResponseType } from 'ofetch'
 import { resolve } from 'pathe'
-import { build, createDevServer, prepare } from 'nitropack'
+import { build, copyPublicAssets, createDevServer, prepare, prerender } from 'nitropack'
+import { listen } from 'listhen'
 import { createTestContext, useTestContext } from './context'
+import type { SetupOptions } from './types'
 
 declare module 'vitest' {
   export interface ProvidedContext {
@@ -44,24 +46,45 @@ export async function $fetch<T = any, R extends ResponseType = 'json'>(
   return response as TestFetchResponse<MappedResponseType<R, T>>
 }
 
-export interface SetupOptions {
-  rootDir?: string
-}
-
+/**
+ * Setup options for the test context.
+ *
+ * @example
+ * import { setup } from 'nitro-test-utils'
+ *
+ * await setup({
+ *  rootDir: fileURLToPath(new URL('fixture', import.meta.url)),
+ * })
+ *
+ */
 export async function setup(options: SetupOptions) {
   const ctx = await createTestContext(options)
 
   const vitest = await import('vitest')
 
   vitest.beforeAll(async () => {
-    const devServer = createDevServer(ctx.nitro)
-    ctx.server = await devServer.listen({})
-    await prepare(ctx.nitro)
-    const ready = new Promise<void>((resolve) => {
-      ctx.nitro!.hooks.hook('dev:reload', () => resolve())
-    })
-    await build(ctx.nitro)
-    await ready
+    // Setup development server
+    if (ctx.isDev) {
+      const devServer = createDevServer(ctx.nitro)
+      ctx.server = await devServer.listen({})
+      await prepare(ctx.nitro)
+      const ready = new Promise<void>((resolve) => {
+        ctx.nitro!.hooks.hook('dev:reload', () => resolve())
+      })
+      await build(ctx.nitro)
+      await ready
+    }
+    // Build and serve production
+    else {
+      await prepare(ctx.nitro)
+      await copyPublicAssets(ctx.nitro)
+      await prerender(ctx.nitro)
+      await build(ctx.nitro)
+
+      const entryPath = resolve(ctx.nitro.options.output.dir, 'server/index.mjs')
+      const { listener } = await import(entryPath)
+      ctx.server = await listen(listener)
+    }
 
     // eslint-disable-next-line no-console
     console.log('> Nitro server running at', ctx.server.url)
