+++ src/types.ts
@@ -7,3 +7,15 @@ export interface NitroTestContext {
   server?: Listener
   isDev: boolean
 }
+
+export interface SetupOptions {
+  /**
+   * The root directory of the Nitro project to test against.
+   *
+   * @example
+   * fileURLToPath(new URL('fixture', import.meta.url))
+   *
+   * @default process.cwd()
+   */
+  rootDir?: string
+}
