+++ vitest.config.ts
@@ -4,8 +4,11 @@ import { defineConfig } from './dist/config.mjs'
 export default defineConfig({
   test: {
     include: ['test/routes.test.ts', 'test/setup.test.ts'],
-  },
-  nitro: {
-    rootDir: 'test/fixture',
+    forceRerunTriggers: [
+      '**/test/fixture/**/*.ts',
+      '**/test/fixture2/**/*.ts',
+
+      '**/src/**/*.ts',
+    ],
   },
 })
